Add first version
This is a suggestion for an API that people can use to integrate with
happo directly through a javascript bundle. There are two main
functions:

  const happoStatic = require('happo-static');
  happoStatic.init();

Called to initialize the happoStatic package. This will set up some
globals used by happo workers. And might be used for other things as
well.

happoStatic.registerExample({
  component: 'Foo',
  variant: 'default',
  render: () => {
    document.body.innerHTML = 'foo bar';
  },
})